requirements.txt
colorzero==2.0
gpiozero==1.6.2
pyserial==3.5
